requirements.txt
fastapi
uvicorn
requests
fpdf
python-dotenv
fpdf
apscheduler